+++ rustconn-core/tests/properties/config_tests.rs
@@ -175,6 +175,7 @@ fn arb_rdp_config() -> impl Strategy<Value = RdpConfig> {
                 scale_override: Default::default(),
                 disable_nla: false,
                 clipboard_enabled: true,
+                show_local_cursor: true,
             },
         )
 }
@@ -216,6 +217,7 @@ fn arb_vnc_config() -> impl Strategy<Value = VncConfig> {
             clipboard_enabled: true,
             custom_args,
             scale_override: Default::default(),
+            show_local_cursor: true,
         })
 }